# src/main/java/com/javbre/dto/AnswerDto.java
package com.javbre.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class AnswerDto {

    @JsonProperty("questionId")
    private Long questionId;
    @JsonProperty("chosenOptionId")
    private Long chosenOptionId;
}