# src/main/java/com/javbre/dto/AssignRequest.java
package com.javbre.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.List;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class AssignRequest {

    @JsonProperty("studentIds")
    private List<Long> studentIds;
}