# src/main/java/com/javbre/dto/OptionDto.java
package com.javbre.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class OptionDto {

    @JsonProperty("text")
    private String text;
    @JsonProperty("isCorrect")
    private boolean isCorrect;
}