# src/main/java/com/javbre/dto/SubmitAnswersRequest.java
package com.javbre.dto;

import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Getter;
import lombok.NoArgsConstructor;
import lombok.Setter;

import java.util.List;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
public class SubmitAnswersRequest {

    @JsonProperty("answers")
    private List<AnswerDto> answers;
}